Revert "Move functions for unprivileged access to API dir"
This reverts commit 0b5921aa68658ec885fba53967608d14d3dd0c3b.
